All security constraints added

+++ contoller/TodoController.js
@@ -63,55 +63,60 @@ exports.insertTodo = (req,res,next) => {
 };
 
 exports.removeAllTodo = (req,res,next) => {
-
     TodoModel.find({})
     .populate('user')
     .then((todo) => {
         if(todo != null && req.user._id != null){
             
-            let data = todo;
-            let flag = false;
-            
-
-            for(let i = 0; i < data.length ;i++){
-                if((data[i].user._id).toString() == (req.user._id).toString())
+            for(let i = 0; i < todo.length ;i++){
+                if((todo[i].user._id).toString() == (req.user._id).toString())
                     {
                         flag = true;
-                        todo[i].pop();
+                        todo[i].remove();
                     }
             }
 
 
             if(flag){
-                           
-                todo.save()
-                .then((todo) => {
-                    res.statusCode = 200;
-                    res.setHeader('Content-Type','application/json');
-                    res.json({status : true , success : 'Todo Deleted Successfull !'});
-                })
-
+                res.statusCode = 200;
+                res.setHeader('Content-Type','application/json');
+                res.json({status : true , msg : "User Data is Deleted Successfully"});   
             }else{
                 res.statusCode = 404;
                 res.setHeader('Content-Type','application/json');
                 res.json({status : true , msg : "User Data not found"});   
             }
 
+        }else{
+            res.statusCode = 404;
+            res.setHeader('Content-Type','application/json');
+            res.json({status : true , msg : "User Data not found"});   
         }
     },(err) => next(err))
     .catch((err) => next(err));
 
-
-
+    
 }
 
 exports.getTodoById = (req,res,next) => {
     TodoModel.findById(req.params.todoId)
+    .populate('user')
     .then((todo) => {
         if(todo != null){
-            res.statusCode = 200;
-            res.setHeader('Content-Type','application/json');
-            res.json(todo);
+
+            let id1 = req.user._id;
+            let id2 = todo.user._id;
+
+            if(id1.toString() === id2.toString()){
+                res.statusCode = 200;
+                res.setHeader('Content-Type','application/json');
+                res.json(todo);
+            }else{
+                res.statusCode = 401;
+                res.setHeader('Content-Type','application/json');
+                res.json({status : true, msg : 'Unauthorized User'});
+            }
+
         }else{
             let err = new Error('Todo '+req.params.todoId+" not found");
             err.status = 404;
@@ -124,23 +129,38 @@ exports.getTodoById = (req,res,next) => {
 
 exports.updateTodoById = (req,res,next) => {
         TodoModel.findById(req.params.todoId)
+        .populate('user')
         .then((todo) => {
             if(todo != null){
 
-               if(req.body.todo){
-                   todo.todo = req.body.todo;
-               }
-               if(req.body.isChecked){
-                   todo.isChecked = req.body.isChecked;
-               }
 
-               todo.save()
-               .then((todo)=>{
-                    res.statusCode = 200;
+                let id1 = req.user._id;
+                let id2 = todo.user._id;
+    
+                if(id1.toString() === id2.toString()){
+
+                    if(req.body.todo){
+                        todo.todo = req.body.todo;
+                    }
+                    if(req.body.isChecked){
+                        todo.isChecked = req.body.isChecked;
+                    }
+     
+                    todo.save()
+                    .then((todo)=>{
+                         res.statusCode = 200;
+                         res.setHeader('Content-Type','application/json');
+                         res.json({success : true,status : "Todo Updated Succesfully !",todo : todo});
+                    },(err) => next(err))
+                    .catch((err) => next(err));
+
+                }else{
+                    res.statusCode = 401;
                     res.setHeader('Content-Type','application/json');
-                    res.json({success : true,status : "Todo Updated Succesfully !",todo : todo});
-               },(err) => next(err))
-               .catch((err) => next(err));
+                    res.json({status : true, msg : 'Unauthorized User'});
+                }
+
+              
 
             }else{
                 let err = new Error('Todo '+req.params.todoId+" not found");
@@ -154,12 +174,33 @@ exports.updateTodoById = (req,res,next) => {
 };
 
 exports.removeTodoById = (req,res,next) =>{
-    TodoModel.findByIdAndRemove(req.params.todoId)
+    TodoModel.findById(req.params.todoId)
+    .populate('user')
     .then((todo) => {
         if(todo != null){
-            res.statusCode = 200;
-            res.setHeader('Content-Type','application/json');
-            res.json({status : true, success : "Todo Deleted Successfully !",todo : todo});
+
+
+            
+            let id1 = req.user._id;
+            let id2 = todo.user._id;
+           
+            
+            if(id1.toString() === id2.toString()){
+            
+                TodoModel.findByIdAndRemove(req.params.todoId)
+                .then((todo)=>{
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type','application/json');
+                    res.json({success : true,status : "Todo Deleted Succesfully !"});
+            
+                },(err) => next(err))
+                .catch((err) => next(err))
+
+            }else{
+                    res.statusCode = 401;
+                    res.setHeader('Content-Type','application/json');
+                    res.json({status : true, msg : 'Unauthorized User'});
+            }
         }else{
             let err = new Error('Todo '+req.params.todoId+" not found");
             err.status = 404;
